# swagger_doc.yml
openapi: 3.0.0
info:
  title: Auto-École API
  description: API pour la gestion d'une application d'auto-école
  version: 1.0.0
servers:
  - url: https://virtserver.swaggerhub.com/monipermis-76d/server/1.0.0
    description: SwaggerHub API Auto Mocking
  - url: http://localhost:8080/api/v1
    description: Serveur de développement local
...
paths:
...
  /instructors/{instructorId}/availability:
    get:
      tags:
        - Disponibilités
        - Instructeurs
      summary: Récupérer les disponibilités d'un instructeur
      parameters:
        - name: instructorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Disponibilités récupérées avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AvailabilitySchedule'
        "404":
          description: Instructeur non trouvé
    post:
      tags:
        - Disponibilités
        - Instructeurs
      summary: Ajouter une disponibilité pour un instructeur
      parameters:
        - name: instructorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityScheduleInput'
        required: true
      responses:
        "201":
          description: Disponibilité ajoutée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilitySchedule'
        "404":
          description: Instructeur non trouvé
  /instructors/{instructorId}/availability/{availabilityId}:
    put:
      tags:
        - Disponibilités
        - Instructeurs
      summary: Modifier une disponibilité pour un instructeur
      parameters:
        - name: instructorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: availabilityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AvailabilityScheduleInput'
        required: true
      responses:
        "200":
          description: Disponibilité mise a jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilitySchedule'
        "404":
          description: Instructeur non trouvé
    delete:
      tags:
        - Disponibilités
        - Instructeurs
      summary: Supprimer une disponibilité pour un instructeur
      parameters:
        - name: instructorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: availabilityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Disponibilité supprimée avec succès
        "404":
          description: Instructeur non trouvé
  /instructors/{instructorId}/unavailability:
    get:
      tags:
        - Disponibilités
        - Instructeurs
      summary: Récupérer les indisponibilités d'un instructeur
      parameters:
        - name: instructorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Indisponibilités récupérées avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InstructorUnavailability'
        "404":
          description: Instructeur non trouvé
    post:
      tags:
        - Disponibilités
        - Instructeurs
      summary: Ajouter une indisponibilité pour un instructeur
      parameters:
        - name: instructorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorUnavailabilityInput'
        required: true
      responses:
        "201":
          description: Indisponibilité ajoutée avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstructorUnavailability'
        "404":
          description: Instructeur non trouvé
  /instructors/{instructorId}/unavailability/{unavailabilityId}:
    put:
      tags:
        - Disponibilités
        - Instructeurs
      summary: Modifier une indisponibilité pour un instructeur
      parameters:
        - name: instructorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: unavailabilityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InstructorUnavailabilityInput'
        required: true
      responses:
        "200":
          description: Indisponibilité mise à jour avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InstructorUnavailability'
        "400":
          description: Requête invalide (ex. données manquantes ou mal formées)
        "401":
          description: Non authentifié – un token valide est requis
        "403":
          description: Accès refusé – vous n’avez pas les droits nécessaires
        "404":
          description: Ressource non trouvée (instructeur ou indisponibilité inexistante)
        "409":
          description: Conflit – l'indisponibilité modifiée entre en conflit avec une autre
        "422":
          description: Données de validation incorrectes
        "500":
          description: Erreur interne du serveur

    delete:
      tags:
        - Disponibilités
        - Instructeurs
      summary: Supprimer une indisponibilité pour un instructeur
      parameters:
        - name: instructorId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: unavailabilityId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Indisponibilité supprimée avec succès
        "404":
          description: Ressource non trouvée (instructeur ou indisponibilité inexistante)
...
components:
  schemas:
...
    AvailabilitySchedule:
      required:
        - endDateTime
        - id
        - instructorId
        - isRecurring
        - startDateTime
      type: object
      properties:
        id:
          type: integer
          format: int64
        instructorId:
          type: integer
          format: int64
        startDateTime:
          type: string
          description: Date et heure de début de la disponibilité
          format: date-time
          example: 2024-01-15T09:00:00Z
        endDateTime:
          type: string
          description: Date et heure de fin de la disponibilité
          format: date-time
          example: 2024-01-15T17:00:00Z
        isRecurring:
          type: boolean
          description: Indique si la disponibilité est récurrente
          example: true
        recurrenceRule:
          type: string
          description: Règle de récurrence (optionnel)
          nullable: true
          example: WEEKLY
          enum:
            - DAILY
            - WEEKLY
            - MONTHLY
            - YEARLY
        expiryDate:
          type: string
          description: Date d'expiration de la récurrence (optionnel)
          format: date
          nullable: true
          example: 2024-12-31
        note:
          maxLength: 255
          type: string
          description: Note additionnelle (optionnel)
          nullable: true
          example: Disponible uniquement le matin
    AvailabilityScheduleInput:
      required:
        - endDateTime
        - id
        - instructorId
        - isRecurring
        - startDateTime
      type: object
      properties:
        startDateTime:
          type: string
          description: Date et heure de début de la disponibilité
          format: date-time
          example: 2024-01-15T09:00:00Z
        endDateTime:
          type: string
          description: Date et heure de fin de la disponibilité
          format: date-time
          example: 2024-01-15T17:00:00Z
        isRecurring:
          type: boolean
          description: Indique si la disponibilité est récurrente
          example: true
        recurrenceRule:
          type: string
          description: Règle de récurrence (optionnel)
          nullable: true
          example: WEEKLY
          enum:
            - DAILY
            - WEEKLY
            - MONTHLY
            - YEARLY
        expiryDate:
          type: string
          description: Date d'expiration de la récurrence (optionnel)
          format: date
          nullable: true
          example: 2024-12-31
        note:
          maxLength: 255
          type: string
          description: Note additionnelle (optionnel)
          nullable: true
          example: Disponible uniquement le matin
    InstructorUnavailability:
      required:
        - endDatetime
        - id
        - instructorId
        - reason
        - startDatetime
      type: object
      properties:
        id:
          type: integer
          format: int64
        instructorId:
          type: integer
          format: int64
        startDatetime:
          type: string
          format: date-time
        endDatetime:
          type: string
          format: date-time
        reason:
          type: string
    InstructorUnavailabilityInput:
      required:
        - endDatetime
        - reason
        - startDatetime
      type: object
      properties:
        startDatetime:
          type: string
          format: date-time
        endDatetime:
          type: string
          format: date-time
        reason:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT